requirements.txt
discord.py
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2